cbdcc now will now use the server if CBD_SERVER is set to find workers
The code will prioritize CBD_POTENTIAL_HOST over CBD_SERVER if present. If
the remote build fails for some reason a local build is run instead.

The test.sh bash script has also been updated to check the output of the
test progam and to test the server version of the compile build.