We now find and reporting workers by IP address instead of hostname
This is more reliable because we cannot always expect hostnames to resolve
properly on different machines.  This is also quite a bit more complex
because hosts can have several IP addresses so we report all available IP
addresses, and the server matches up the most "local" relative to each
client.

We have also included more debug logging to make it easier to debug
connection issues in the future.