Added an autodiscovery logic
The server listens for UDP broadcasts from the client, and responds with
a direct UDP packet containing the port the client can connect on. This is
intended be used to ease the configuration pain.  This work is not yet
integrated into the main program, and doesn't support any notion of
sessions.