Generate a unique machine ID based on MAC address
We don't want to rely on host names, which aren't guaranteed to be
unique, to identify nodes.  The gets all the MAC addresses in string
form then picks the first one in sorted order.  As long as the hardware
in a machine does not change it will always have the same ID.