Support calling program through cbd-program symlink
In this mode cbd reads the program it should be invoking by parsing the
name the program is given.  This allows for integration with software
that needs to given just a path to compiler binary.

The unit tests have been updated, and the build script adds symlinks for
gcc & g++.